third_party/blink/web_tests/platform/generic/fast/filesystem/simple-readonly-expected.txt

FileSystem API readonly attributes test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

webkitRequestFileSystem(TEMPORARY, 100, successCallback, errorCallback)
trying to set readonly property fileSystem.name
fileSystem.name = 'bar'
PASS fileSystem.name is still file__0:Temporary
root = fileSystem.root
root.getFile('foo', {create:true}, getFileCallback, errorCallback)
trying to set readonly property fileEntry.isFile
fileEntry.isFile = false
PASS fileEntry.isFile is still true
trying to set readonly property fileEntry.isDirectory
fileEntry.isDirectory = true
PASS fileEntry.isDirectory is still false
trying to set readonly property fileEntry.name
fileEntry.name = 'bar'
PASS fileEntry.name is still foo
trying to set readonly property fileEntry.fullPath
fileEntry.fullPath = 'bar'
PASS fileEntry.fullPath is still /foo
trying to set readonly property fileEntry.filesystem
fileEntry.filesystem = null
PASS fileEntry.filesystem is still [object DOMFileSystem]
PASS successfullyParsed is true

TEST COMPLETE
